Refactor volatility helpers
